refactor: Move common types
