+++ packages/tools/client-debugger/client-debugger/src/data-visualization/index.ts
@@ -23,7 +23,6 @@ export {
 export {
 	createHandleNode,
 	FluidHandleNode,
-	FluidObjectId,
 	FluidObjectNode,
 	FluidObjectNodeBase,
 	FluidObjectTreeNode,
